+++ code/packages/store-azure-web-pubsub/src/client.ts
@@ -25,6 +25,7 @@ import {
 import type { WeaveStoreAzureWebPubsub } from './store-azure-web-pubsub';
 import { WEAVE_STORE_CONNECTION_STATUS } from '@inditextech/weave-types';
 import { WEAVE_STORE_AZURE_WEB_PUBSUB_CONNECTION_STATUS } from './constants';
+import { handleChunkedMessage, handleMessageBufferData } from './utils';
 
 const messageSyncStep1 = 0;
 const messageAwareness = 1;
@@ -119,6 +120,7 @@ export class WeaveStoreAzureWebPubSubSyncClient extends Emittery {
   private _awareness!: awarenessProtocol.Awareness;
   private _options: WeaveStoreAzureWebPubSubSyncClientOptions;
   private _initialized: boolean;
+  private _chunkedMessages: Map<string, string[]>;
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private _updateHandler: (update: any, origin: any) => void;
@@ -161,6 +163,7 @@ export class WeaveStoreAzureWebPubSubSyncClient extends Emittery {
     this._status = WEAVE_STORE_AZURE_WEB_PUBSUB_CONNECTION_STATUS.DISCONNECTED;
     this._wsConnected = false;
     this._initialized = false;
+    this._chunkedMessages = new Map();
 
     this._connectionRetries = 0;
     this._synced = false;
@@ -272,13 +275,13 @@ export class WeaveStoreAzureWebPubSubSyncClient extends Emittery {
         }
 
         const resyncInSeconds = this._options.resyncInterval / 1000;
-        const resyncLimitInSeconds = resyncInSeconds + 5;
+        const resyncWindowLimitInSeconds = resyncInSeconds + 10; // twice size window and 10 seconds grace period
         const now = new Date();
         const diffInSeconds =
           (now.getTime() - this._lastReceivedSyncResponse) / 1000;
 
         if (
-          diffInSeconds > resyncLimitInSeconds &&
+          diffInSeconds > resyncWindowLimitInSeconds &&
           this._ws?.readyState === WebSocket.OPEN
         ) {
           // if last received sync response is older than resync interval + 5s, assume connection lost
@@ -462,8 +465,27 @@ export class WeaveStoreAzureWebPubSubSyncClient extends Emittery {
         return;
       }
 
-      const buf = Buffer.from(messageData.c, 'base64');
-      const encoder = readMessage(this, buf, true, messageData.f);
+      const joinedMessagePayload = handleChunkedMessage(
+        this._chunkedMessages,
+        messageData
+      );
+
+      if (messageData.type === 'chunk') {
+        // skip processed chunked message
+        return;
+      }
+
+      const buffer = handleMessageBufferData(
+        messageData.c,
+        joinedMessagePayload
+      );
+
+      if (!buffer) {
+        // no buffer found, ignore message
+        return;
+      }
+
+      const encoder = readMessage(this, buffer, true, messageData.f);
       if (encoding.length(encoder) > 1) {
         sendToControlGroup(
           this,
@@ -578,9 +600,6 @@ function safeSend(data: string) {
   const bytes = new TextEncoder().encode(data);
 
   if (bytes.byteLength > MAX_BYTES) {
-    console.warn(
-      `Message too large: ${bytes.byteLength} bytes (limit ${MAX_BYTES}). Skipping send.`
-    );
     return false;
   }
 
@@ -609,6 +628,7 @@ function sendToControlGroup(
   const payload = JSON.stringify({
     type: MessageType.SendToGroup,
     group: `${group}.host`,
+    noEcho: true,
     data: {
       t: type,
       f: client.id,
@@ -617,8 +637,63 @@ function sendToControlGroup(
   });
 
   if (!safeSend(payload)) {
+    sendToControlGroupChunked(client, group, type, u8);
     return;
   }
+
+  client.ws?.send(payload);
+}
+
+function chunkString(str: string, size: number) {
+  const chunks = [];
+  for (let i = 0; i < str.length; i += size) {
+    chunks.push(str.slice(i, i + size));
+  }
+  return chunks;
+}
+
+function sendToControlGroupChunked(
+  client: WeaveStoreAzureWebPubSubSyncClient,
+  group: string,
+  type: string,
+  u8: Uint8Array
+) {
+  const base64Data = Buffer.from(u8).toString('base64');
+
+  const CHUNK_SIZE = 60 * 1024; // 60 KB
+  const chunks = chunkString(base64Data, CHUNK_SIZE);
+  const payloadId = uuidv4();
+
+  for (let i = 0; i < chunks.length; i++) {
+    const payload = JSON.stringify({
+      type: MessageType.SendToGroup,
+      group: `${group}.host`,
+      noEcho: true,
+      data: {
+        payloadId,
+        type: 'chunk',
+        totalChunks: chunks.length,
+        index: i,
+        t: type,
+        f: client.id,
+        c: chunks[i],
+      },
+    });
+
+    client.ws?.send(payload);
+  }
+
+  const payload = JSON.stringify({
+    type: MessageType.SendToGroup,
+    group: `${group}.host`,
+    noEcho: true,
+    data: {
+      payloadId,
+      type: 'end',
+      f: client.id,
+      t: type,
+    },
+  });
 
   client.ws?.send(payload);
 }